model/requirements.txt
requests
spacy
python-dotenv
en_core_web_sm
